feat: move edge metrics regression tests to tests folder
